tests: Fold distributed-engine unit tests into test_data.py
These are serial unit tests of devito/data internals (Selection, Layout,
result_dims, index_has_array), same as TestDecomposition, so they belong
beside it rather than in a separate file. Move the four classes into
test_data.py (aliasing the engine's Scalar to avoid clashing with
devito.types.Scalar) and drop tests/test_distributed_engine.py.
